06_TradeVision/src/trade_vision_bot/requirements.txt
pyzmq==25.1.2
supabase==2.3.0
python-dotenv==1.0.0
requests==2.31.0
